# Conflicts:
#	views/authed.twig

+++ lib/Listener/OnUserDelete/MediaListener.php
@@ -1,5 +1,24 @@
 <?php
-
+/**
+ * Copyright (C) 2021 Xibo Signage Ltd
+ *
+ * Xibo - Digital Signage - http://www.xibo.org.uk
+ *
+ * This file is part of Xibo.
+ *
+ * Xibo is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * Xibo is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with Xibo.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 namespace Xibo\Listener\OnUserDelete;
 
@@ -44,11 +63,12 @@ class MediaListener implements OnUserDeleteInterface
         $user = $event->getUser();
         $function = $event->getFunction();
         $newUser = $event->getNewUser();
+        $systemUser = $event->getSystemUser();
 
         if ($function === 'delete') {
-            $this->deleteChildren($user, $dispatcher);
+            $this->deleteChildren($user, $dispatcher, $systemUser);
         } elseif ($function === 'reassignAll') {
-            $this->reassignAllTo($user, $newUser);
+            $this->reassignAllTo($user, $newUser, $systemUser);
         } elseif ($function === 'countChildren') {
             $event->setReturnValue($event->getReturnValue() + $this->countChildren($user));
         }
@@ -57,10 +77,10 @@ class MediaListener implements OnUserDeleteInterface
     /**
      * @inheritDoc
      */
-    public function deleteChildren(User $user, EventDispatcherInterface $dispatcher)
+    public function deleteChildren(User $user, EventDispatcherInterface $dispatcher, User $systemUser)
     {
         // Delete any media
-        foreach ($this->mediaFactory->getByOwnerId($user->userId) as $media) {
+        foreach ($this->mediaFactory->getByOwnerId($user->userId, 1) as $media) {
             // If there is a parent, bring it back
             try {
                 $parentMedia = $this->mediaFactory->getParentById($media->mediaId);
@@ -71,21 +91,27 @@ class MediaListener implements OnUserDeleteInterface
                 // This is fine, no parent
                 $parentMedia = null;
             }
-            $dispatcher->dispatch(MediaDeleteEvent::$NAME, new MediaDeleteEvent($media, $parentMedia));
-            $media->delete();
+
+            // if this User owns any module files, reassign to systemUser instead of deleting.
+            if ($media->mediaType === 'module') {
+                $media->setOwner($systemUser->userId);
+                $media->save();
+            } else {
+                $dispatcher->dispatch(MediaDeleteEvent::$NAME, new MediaDeleteEvent($media, $parentMedia, true));
+                $media->delete();
+            }
         }
     }
 
     /**
      * @inheritDoc
      */
-    public function reassignAllTo(User $user, User $newUser)
+    public function reassignAllTo(User $user, User $newUser, User $systemUser)
     {
-        // Reassign media
-        $this->storageService->update('UPDATE `media` SET userId = :userId WHERE userId = :oldUserId', [
-            'userId' => $newUser->userId,
-            'oldUserId' => $user->userId
-        ]);
+        foreach ($this->mediaFactory->getByOwnerId($user->userId, 1) as $media) {
+            ($media->mediaType === 'module') ? $media->setOwner($systemUser->userId) : $media->setOwner($newUser->getOwnerId());
+            $media->save();
+        }
     }
 
     /**
@@ -93,9 +119,9 @@ class MediaListener implements OnUserDeleteInterface
      */
     public function countChildren(User $user)
     {
-        $media = $this->mediaFactory->getByOwnerId($user->userId);
+        $media = $this->mediaFactory->getByOwnerId($user->userId, 1);
         $count = count($media);
-        $this->getLogger()->debug(sprintf('Counted Children Media on %d, there are %d', $user->userId, $count));
+        $this->getLogger()->debug(sprintf('Counted Children Media on User ID %d, there are %d', $user->userId, $count));
 
         return $count;
     }
